Reorganize module dependencies
