+++ langextract/__init__.py
@@ -66,7 +66,7 @@ _LAZY_MODULES = {
     "chunking": "langextract.chunking",
     "data": "langextract.data",
     "data_lib": "langextract.data_lib",
-    "debug_utils": "langextract.debug_utils",
+    "debug_utils": "langextract.core.debug_utils",
     "exceptions": "langextract.exceptions",
     "factory": "langextract.factory",
     "inference": "langextract.inference",
